requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
Pillow>=8.0.0
scikit-learn>=0.24.0
numpy>=1.21.0
tqdm>=4.65.0
pandas>=1.5.0
chardet>=4.0.0 